# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DoTheSplit API
  version: 0.2.0
  description: |
    Split expenses among group members.


    Expense-sharing API for the DoTheSplit app.

    Conventions:
      * All money values are integer cents (no floats).
      * All IDs are UUIDs.
      * Authentication is an opaque session cookie (`dts_session` on plain HTTP,
        `__Host-dts_session` on HTTPS) set by `/v1/auth/login`.
      * The session cookie is `SameSite=Lax`; write endpoints rely on that for CSRF.
      * Request bodies reject unknown fields (`additionalProperties: false`).
      * Times are RFC 3339 UTC.

    Versioning:
      * Business endpoints live under `/v1`. Breaking changes → `/v2`.
      * `/healthz` and `/readyz` are infrastructure endpoints and are NOT versioned.
  contact:
    name: DoTheSplit
    url: https://github.com/julian-alarcon/dothesplit
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:8080
    description: Local dev (Docker Compose)
...
security:
  - cookieAuth: []
paths:
...
  /v1/auth/verify:
    post:
      summary: Verify a pending registration with the 6-digit code
      operationId: verifyEmail
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyEmailRequest" }
      responses:
        "200":
          description: Verified; session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "410":
          description: Verification code expired or already used.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/verify/resend:
    post:
      summary: Resend the registration verification code
      description: |
        Always returns 204 regardless of whether the email matches a pending
        registration, to avoid account enumeration. Rate-limited per IP.
      operationId: resendVerification
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResendVerificationRequest" }
      responses:
        "204":
          description: Accepted. A new code may have been sent.
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /v1/me:
    get:
      summary: Get the current user
      operationId: me
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      summary: Rename the current user
      operationId: updateMe
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMeRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      summary: Soft-delete the current user (tombstones display_name, scrubs email, clears all sessions)
      operationId: deleteMe
      tags: [auth]
      responses:
        "204":
          description: Account deleted; session cleared
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/me/password:
    post:
      summary: Change the current user's password (requires old password)
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "204":
          description: Password changed
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/me/email/change-request:
    post:
      summary: Begin a change-email flow (sends a code to the new address)
      description: |
        Step-up: the caller's current password must be supplied. The new email
        is held in a verification token and only persisted to `users.email`
        when the user POSTs the matching 6-digit code to
        `/v1/me/email/change-confirm`. The current address keeps working until
        confirm succeeds.
      operationId: changeEmailRequest
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeEmailRequest" }
      responses:
        "202":
          description: Code sent (if SMTP is configured); awaiting confirmation.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "423":
          description: Step-up rate-limited.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/me/email/change-confirm:
    post:
      summary: Confirm a pending email change with the 6-digit code
      description: |
        On success the user's email is updated, all sessions for the user are
        revoked, and a fresh session cookie is issued so the current browser
        stays logged in.
      operationId: changeEmailConfirm
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfirmEmailChangeRequest" }
      responses:
        "200":
          description: Email updated; new session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410":
          description: Verification code expired or already used.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/me/notifications:
    get:
      summary: Get the current user's notification preferences
      operationId: getMyNotifications
      tags: [auth]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPrefs" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      summary: Update the current user's notification preferences
      operationId: updateMyNotifications
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationPrefs" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPrefs" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/me/avatar:
    put:
      summary: Upload an 8x8 avatar PNG (base64-encoded, <= 1024 bytes)
      operationId: setAvatar
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetAvatarRequest" }
      responses:
        "204":
          description: Avatar saved
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      summary: Remove the current user's avatar (falls back to initials on display)
      operationId: deleteAvatar
      tags: [auth]
      responses:
        "204":
          description: Avatar cleared
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/users/{id}/avatar:
    get:
      summary: Serve a user's pixelated avatar PNG (group-member access)
      description: |
        Returns the stored avatar PNG for the user. The source is an 8x8 bitmap
        uploaded by the client; the server upscales it with nearest-neighbour
        sampling before storage so it renders crisp at any CSS size. Callers
        must share at least one group with the target user. Browsers should
        cache on `?v={avatar_updated_at}` to avoid stale renders when the owner
        updates.
      operationId: getUserAvatar
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: PNG bytes
          content:
            image/png:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dts_session
      description: |
        Session cookie. Name is `dts_session` on plain HTTP deployments and
        `__Host-dts_session` when `COOKIE_SECURE=true`.
...
  responses:
    BadRequest:
      description: Invalid input (malformed JSON, failed validation, or unknown field).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid session cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted for this resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist (or is hidden by soft-delete).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Request conflicts with existing state (e.g. duplicate email).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limited. Retry after the `Retry-After` header.
      headers:
        Retry-After:
          description: Seconds until the client may retry.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    User:
      type: object
      required: [id, email, display_name, created_at, has_avatar, week_start]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
        has_avatar:
          type: boolean
          description: True when the user has uploaded a pixel avatar. Fetch it at `/v1/users/{id}/avatar`.
        avatar_updated_at:
          type: string
          format: date-time
          description: Cache-buster for the avatar URL. Changes on every upload/delete.
        deleted_at:
          type: string
          format: date-time
          description: Set when the account has been soft-deleted; `display_name` is then a tombstone.
        week_start:
          type: integer
          enum: [0, 1]
          description: First day of the week in calendar UIs. 0 = Sunday, 1 = Monday.
        timezone:
          type: string
          nullable: true
          description: IANA timezone (e.g. "Europe/Madrid"). Null means the client should use the device-detected zone.
        is_admin:
          type: boolean
          readOnly: true
          description: True when the user has the admin role. Read-only; promotion only happens via the bootstrap path or an admin endpoint.
        must_change_password:
          type: boolean
          readOnly: true
          description: True when an admin reset the user's password. The frontend must redirect them to a password-change page; the API rejects all non-password endpoints until cleared.
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          description: When the email address was confirmed. Null until the user verifies, or set automatically at registration when SMTP is unconfigured.
    UpdateMeRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        display_name: { type: string, minLength: 1, maxLength: 80 }
        week_start:
          type: integer
          enum: [0, 1]
          description: First day of the week. 0 = Sunday, 1 = Monday.
        timezone:
          type: string
          maxLength: 64
          description: IANA timezone (e.g. "Europe/Madrid"). Send an empty string to revert to device detection.
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [old_password, new_password]
      properties:
        old_password: { type: string, minLength: 1, maxLength: 256 }
        new_password: { type: string, minLength: 10, maxLength: 256 }
    SetAvatarRequest:
      type: object
      additionalProperties: false
      required: [png_base64]
      properties:
        png_base64:
          type: string
          minLength: 1
          maxLength: 1500
          description: Base64-encoded PNG. Decoded size must be <= 1024 bytes and the image exactly 8x8 pixels.
...
    VerifyEmailRequest:
      type: object
      additionalProperties: false
      required: [email, code]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: "^[0-9]{6}$"
          description: 6-digit numeric code received by email.
      example:
        email: alice@example.com
        code: "482193"
    ResendVerificationRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: { type: string, format: email, maxLength: 254 }
      example:
        email: alice@example.com
    ChangeEmailRequest:
      type: object
      additionalProperties: false
      required: [new_email, password]
      properties:
        new_email: { type: string, format: email, maxLength: 254 }
        password:
          type: string
          minLength: 1
          description: The caller's current password (step-up).
      example:
        new_email: alice.new@example.com
        password: correct-horse-battery-staple
    ConfirmEmailChangeRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: "^[0-9]{6}$"
          description: 6-digit code received at the new address.
      example:
        code: "917544"
    NotificationPrefs:
      type: object
      additionalProperties: false
      description: |
        Per-user, per-event email opt-in flags. All default to false (off).
        The server only emails when the corresponding flag is true AND SMTP is
        configured.
      properties:
        notify_recurring_run:
          type: boolean
          description: Email me when a recurring expense is materialized in one of my groups.
        notify_settlement:
          type: boolean
          description: Email me when a settlement is recorded in one of my groups.
        notify_group_added:
          type: boolean
          description: Email me when I am added to a new group.
...